requirements.txt
python-dotenv
requests
Telethon
tzlocal
websockets==11.0